reporting/requirements.txt
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.163.0
anthropic==0.49.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.5
pytest-mock==3.14.0
